requirements.txt
# Core dependencies
python-telegram-bot==20.7
aiohttp==3.9.1
websockets==12.0
pandas==2.1.4
python-dotenv==1.0.0
requests==2.31.0

# Data handling and validation
pydantic==2.5.2
numpy==1.26.2

# Database (for bonus features)
aiosqlite==0.19.0

# Async support
asyncio-mqtt==0.16.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging and monitoring
structlog==23.2.0

# Development
jupyter==1.0.0
ipython==8.18.1
